+++ testing/UniformDiscreteTesting.cpp
@@ -1,5 +1,5 @@
-#include "distributions/BasicRandGenerator.hpp"
-#include "distributions/UniformDiscreteRand.hpp"
+#include "RandLib/distributions/BasicRandGenerator.hpp"
+#include "RandLib/distributions/UniformDiscreteRand.hpp"
 
 namespace
 {
